Adds black formatting changes along with some changes to simulation.

+++ src/am/workspace/simulation/utils.py
@@ -1,28 +1,56 @@
 import os
 import pickle
 
+
 class WorkspaceSimulationUtils:
     """
     Utility functions for workspace simulation class.
     """
-    def create_simulation_folder(self, simulation):
+
+    def create_simulation_folders(self, simulation, save_simulation=False):
         """
-        Creates folder for `simulation` and parent `simulations` folder if needed.
-        Also saves passed simulation into `simulation.pkl` within created folder.
+        Creates file and folders for `simulation`.
+        Also creates parent `simulations` folder if not existant.
         """
-        # Creates `simulations` folder path within workspace if not created.
-        simulations_path = os.path.join(self.workspace_path, "simulations")
+
+        # Should happen just once for each workspace initialization.
+        simulations_path = os.path.join(
+            # `/out/<self.workspace_path>/simulations/`
+            self.workspace_path,
+            "simulations",
+        )
         if not os.path.isdir(simulations_path):
             os.makedirs(simulations_path)
 
-        # Create `simulation` folder within workspace path 
-        simulation_path = os.path.join(self.workspace_path, "simulations", simulation.filename)
+        # Creates folder for housing simulation layers and visualizations.
+        simulation_path = os.path.join(
+            # `/out/<self.workspace_path>/simulations/<simulation.filename>/`
+            self.workspace_path,
+            "simulations",
+            simulation.filename,
+        )
         if not os.path.isdir(simulation_path):
             os.makedirs(simulation_path)
 
-        # Save simulation class object to pickle file
-        simulation_pkl_path = os.path.join(simulation_path, "simulation.pkl")
-        with open(simulation_pkl_path, "wb") as file:
-            pickle.dump(simulation, file)
-        
+        # Creates folder for simulation layers.
+        simulation_layers_path = os.path.join(
+            # `/out/<self.workspace_path>/simulations/<simulation.filename>/layers/`
+            self.workspace_path,
+            "simulations",
+            simulation.filename,
+            "layers",
+        )
+        if not os.path.isdir(simulation_layers_path):
+            os.makedirs(simulation_layers_path)
+
+        if save_simulation:
+            # Creates simulation pickle file.
+            simulation_pkl_path = os.path.join(
+                # `/out/<self.workspace_path>/simulations/<simulation.filename>/simulation.pkl`
+                simulation_path,
+                "simulation.pkl",
+            )
+            with open(simulation_pkl_path, "wb") as file:
+                pickle.dump(simulation, file)
+
         return simulation_path
